requirements.txt
psycopg2-binary
django
django-cors-headers
django-filter
djangorestframework
requests
celery
redis
pillow